+++ src/render/mod.rs
@@ -18,7 +18,9 @@ use bevy::{
         render_resource::*,
         renderer::{RenderContext, RenderDevice, RenderQueue},
         texture::{BevyDefault, Image},
-        view::{ComputedVisibility, ExtractedView, ViewUniform, ViewUniformOffset, ViewUniforms},
+        view::{
+            ComputedVisibility, ExtractedView, Msaa, ViewUniform, ViewUniformOffset, ViewUniforms,
+        },
         Extract,
     },
     time::Time,
@@ -725,6 +727,8 @@ pub(crate) struct ParticleRenderPipelineKey {
     /// feature.
     #[cfg(all(feature = "2d", feature = "3d"))]
     pipeline_mode: PipelineMode,
+    /// MSAA sample count.
+    msaa_samples: u32,
 }
 
 impl Default for ParticleRenderPipelineKey {
@@ -734,6 +738,7 @@ impl Default for ParticleRenderPipelineKey {
             particle_texture: None,
             #[cfg(all(feature = "2d", feature = "3d"))]
             pipeline_mode: PipelineMode::Camera3d,
+            msaa_samples: Msaa::default().samples,
         }
     }
 }
@@ -855,7 +860,7 @@ impl SpecializedRenderPipeline for ParticlesRenderPipeline {
             },
             depth_stencil,
             multisample: MultisampleState {
-                count: 4, // TODO: Res<Msaa>.samples
+                count: key.msaa_samples,
                 mask: !0,
                 alpha_to_coverage_enabled: false,
             },
@@ -1485,7 +1490,7 @@ pub(crate) fn prepare_effects(
         .effects
         .iter()
         .map(|(entity, extracted_effect)| {
-            let id = effects_meta.entity_map.get(entity).unwrap().clone();
+            let id = *effects_meta.entity_map.get(entity).unwrap();
             let slice = effects_meta.effect_cache.get_slice(id);
             (slice, extracted_effect, id.0 as u32)
         })
@@ -1686,7 +1691,6 @@ pub(crate) fn prepare_effects(
             force_field: extracted_force_field, // extracted_effect.force_field,
             seed: random::<u32>(),
             effect_index,
-            ..Default::default()
         };
         trace!("spawner_params = {:?}", spawner_params);
         effects_meta.spawner_buffer.push(spawner_params);
@@ -1833,6 +1837,7 @@ pub(crate) fn queue_effects(
     effect_batches: Query<(Entity, &mut EffectBatch)>,
     events: Res<EffectAssetEvents>,
     read_params: QueueEffectsReadOnlyParams,
+    msaa: Res<Msaa>,
 ) {
     trace!("queue_effects");
 
@@ -2150,6 +2155,7 @@ pub(crate) fn queue_effects(
                         shader: batch.render_shader.clone(),
                         #[cfg(feature = "3d")]
                         pipeline_mode: PipelineMode::Camera2d,
+                        msaa_samples: msaa.samples,
                     },
                 );
                 trace!("Render pipeline specialized: id={:?}", render_pipeline_id);
@@ -2251,6 +2257,7 @@ pub(crate) fn queue_effects(
                         shader: batch.render_shader.clone(),
                         #[cfg(feature = "2d")]
                         pipeline_mode: PipelineMode::Camera3d,
+                        msaa_samples: msaa.samples,
                     },
                 );
                 trace!("Render pipeline specialized: id={:?}", render_pipeline_id);